testing/requirements.txt
Flask>=2.0
requests>=2.28
python-dotenv>=1.0
flask-cors>=3.0
python-dateutil>=2.8
Flask==2.3.2
python-dotenv==1.0.0
requests==2.31.0